% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rusterize.r
\name{rusterize}
\alias{rusterize}
\title{Extremely fast rasterization engine built in Rust for R}
\usage{
rusterize(
  data,
  like = NULL,
  resolution = NULL,
  out_shape = NULL,
  extent = NULL,
  field = NULL,
  by = NULL,
  burn = NULL,
  fun = "last",
  background = NA,
  encoding = "dense",
  all_touched = FALSE,
  tap = FALSE,
  dtype = "double"
)
}
\arguments{
\item{data}{An \code{sf} or \code{sfc} object of geometries.}

\item{like}{A \code{terra} template used as a spatial blueprint (resolution, dimension, extent).
Mutually exclusive with \code{resolution}, \code{out_shape}, and \code{extent}.}

\item{resolution}{Pixel resolution as \code{c(xres, yres)} (or a list).}

\item{out_shape}{Output raster dimensions as \code{c(nrows, ncols)} (or a list).}

\item{extent}{Spatial bounding box as \code{c(xmin, ymin, xmax, ymax)} (or a list).}

\item{field}{Column name to use for pixel values. Mutually exclusive with \code{burn}.}

\item{by}{Column used for grouping. Each group is rasterized into a distinct band in the output.}

\item{burn}{A static value or a list of values to apply to each geometries. If a vector, it must match the
length of the geometry data. Mutually exclusive with \code{field}. If a vector, its dtype should match the output dtype,
else it will be internally casted.}

\item{fun}{Pixel function to use when burning geometries. Available options: \code{sum}, \code{first}, \code{last}, \code{min}, \code{max},
\code{count}, or \code{any}.}

\item{background}{Value assigned to pixels not covered by any geometry. Defaults to NA.}

\item{encoding}{The format of the returned object: \code{"dense"} or \code{"sparse"}.}

\item{all_touched}{If True, every pixel touched by a geometry is burned.}

\item{tap}{Target Aligned Pixels: aligns the extent to the pixel resolution.}

\item{dtype}{Output data type: \code{"integer"} or \code{"double"}.}
}
\value{
A \code{terra} object or a \code{SparseArray} in COOrdinate format.
}
\description{
Extremely fast rasterization engine built in Rust for R
}
